Create README.md
Digital nomad portrayed who is a professional working solely in an online environment. while he is leading a location independent and often travel reliant lifestyle where the boundaries between work, leisure and travel appear blurred. Adel aims to conceptualize the digital nomad phenomenon by establishing a definition of digital nomads all across the globe. Further more, he explores  motivations for adapting this lifestyle to others. Perceives work not as an imposed obligation but regard it – much as his leisure activities – as intrinsically motivated and fulfilling, Although crucial for a positive perception of this lifestyle,  exclusive to the silent, faceless elite. He is the whisper behind the company, the beta behind tester, he is the socially engineered end product of Carbon AI.